refactor: ui layout suitable for mobile aspect ratio
