+++ packages/gamev2/src/ui/components/ActionBar.module.css
@@ -1,0 +1,6 @@
+.container {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  margin-left: auto;
+}
