+++ packages/gamev2/src/ui/components/Frame.tsx
@@ -9,12 +9,16 @@ export const Header = ({ children }: PropsWithChildren) => (
   <div className={styles.header}>{children}</div>
 );
 
-export const SidebarContainer = ({ children }: PropsWithChildren) => (
-  <div className={styles.sidebarContainer}>{children}</div>
+export const LeftColumn = ({ children }: PropsWithChildren) => (
+  <div className={styles.leftCol}>{children}</div>
 );
 
-export const Sidebar = ({ children }: PropsWithChildren) => (
-  <div className={styles.sidebar}>{children}</div>
+export const RightColumn = ({ children }: PropsWithChildren) => (
+  <div className={styles.rightCol}>{children}</div>
+);
+
+export const SelectedSlot = ({ children }: PropsWithChildren) => (
+  <div className={styles.selectedSlot}>{children}</div>
 );
 
 export const Footer = ({ children }: PropsWithChildren) => (
